Rename monadic constructs
