+++ crates/cgp-dispatch/src/providers/dispatchers/dispatch_matchers.rs
@@ -1,4 +1,4 @@
-use cgp_monad::monads::ok::OkMonadic;
+use cgp_monad::monadic::ok::OkMonadic;
 use cgp_monad::providers::PipeMonadic;
 
 pub type DispatchMatchers<Providers> = PipeMonadic<OkMonadic, Providers>;
