+++ crates/cgp-monad/src/monadic/mod.rs
