+++ crates/cgp-monad/src/traits/compose.rs
@@ -1,0 +1,3 @@
+pub trait Compose<ProviderA, ProviderB> {
+    type Provider;
+}
