--- crates/cgp-monad/src/traits/monadic.rs
@@ -1,3 +1,0 @@
-pub trait MonadicBind<ProviderA, ProviderB> {
-    type Provider;
-}
